feat: Update kolorsai plugin command filters
Update the command filters for the kolorsai plugin to include "sd" and "dif" in addition to "kolors". This change allows users to trigger the plugin using these additional commands. The update also includes a change in the reply text to "Generating Image..." to provide more accurate feedback to the user.

Refactor the kolors.py utility module

Refactor the negative_prompt variable in the KolorsClient class of the kolors.py utility module. The variable now includes a list of negative prompts for generating images. This change improves the quality and variety of generated images. Steps=25, Scale=7